add react notifications for logging in
