+++ frontend/src/App.js
@@ -1,6 +1,6 @@
 import logo from './logo.svg';
 import './App.css';
-import React, { useState } from 'react';
+import React, { useState, useEffect } from 'react';
 import { Provider } from 'react-redux';
 import {
   BrowserRouter as Router,
@@ -8,14 +8,17 @@ import {
   Route,
   Link
 } from "react-router-dom";
+import {NotificationContainer, NotificationManager} from 'react-notifications';
 
 import store from './store';
 import AuthService from './services/auth.service';
-import LoginView from './components/LoginView'
-import SignupView from './components/SignupView'
+import LoginView from './components/Authentication/LoginView'
+import SignupView from './/components/Authentication/SignupView'
 import HomeView from './components/HomeView'
 import PantryAdminView from './components/PantryAdminView';
 import Navigation from './components/Navigation';
+
+import 'react-notifications/lib/notifications.css';
 
 function App(props) {
   const [username, setUsername] = useState("");
@@ -36,16 +39,57 @@ function App(props) {
         if (response.message){
           // When the API returns `message`,
           // that means the login has failed
-          alert(response.message);
+          createNotification("error", response.message)();
           return -1
         } else {
           setUsername(response.username);
           setToken(response.token);
           setProfile(response.profile);
-          return 0
+
+          createNotification("success", "Successfully logged in!")();
+
+          return 0;
         }
       })
   }
+
+  const createNotification = (type, message) => {
+    console.log("INSIDE CREATE");
+    return () => {
+      // FOR REFERENCE
+      // switch (type) {
+      //   case 'info':
+      //     NotificationManager.info('Info message');
+      //     break;
+      //   case 'success':
+      //     NotificationManager.success('Success message', 'Title here');
+      //     break;
+      //   case 'warning':
+      //     NotificationManager.warning('Warning message', 'Close after 3000ms', 3000);
+      //     break;
+      //   case 'error':
+      //     NotificationManager.error('Error message', 'Click me!', 5000, () => {
+      //       alert('callback');
+      //     });
+      //     break;
+      // }
+      switch (type) {
+        case 'info':
+          // Message, title, timeout
+          NotificationManager.info(message, '', 3000);
+          break;
+        case 'success':
+          NotificationManager.success(message, '', 3000);
+          break;
+        case 'warning':
+          NotificationManager.warning(message, '', 3000);
+          break;
+        case 'error':
+          NotificationManager.error(message, '', 3000)
+          break;
+      }
+    };
+  };
 
   return (
     <Provider store={store}>
@@ -55,13 +99,14 @@ function App(props) {
             <Navigation
               profile={profile}
             />
-
+          <NotificationContainer />
             {/* A <Switch> looks through its children <Route>s and
                 renders the first one that matches the current URL. */}
             <Switch>
               <Route path="/login">
                 <LoginView
                   login={login}
+                  createNotification={createNotification}
                 />
               </Route>
               <Route path="/signup">
