+++ frontend/src/components/Authentication/LoginView.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
 import { Col, Form, Button, Container } from "react-bootstrap";
-import "../../css/LoginView.css";
-import { Link } from "react-router-dom";
+import { Redirect } from "react-router-dom";
 
 class LoginView extends Component {
   constructor(props) {
@@ -9,24 +8,35 @@ class LoginView extends Component {
     this.state = {
       username: "",
       password: "",
+      toHomeView: "", // used for redirection on login success
     };
     this.submitForm = this.submitForm.bind(this);
   }
 
-  submitForm() {
+  async submitForm() {
     const user = {
       username: this.state.username,
       password: this.state.password,
     };
 
-    if (this.state.username.length === 0 || this.state.password.length === 0) {
-      alert("Username or password field is empty.");
+    if (this.state.username.length == 0 || this.state.password.length == 0) {
+      await this.props.createNotification("error", "Username or password is empty.")();
       return;
     }
-    this.props.login(user);
+
+    let loginResult = await this.props.login(user);
+    if (loginResult === 0){
+      this.setState({
+        toHomeView: true
+      })
+    }
   }
 
   render() {
+    if (this.state.toHomeView === true) {
+      return <Redirect to='/' />
+    }
+
     return (
       <Container className="center">
         <Form>
@@ -66,19 +76,13 @@ class LoginView extends Component {
               />
             </Form.Group>
           </Form.Row>
-          <Link
-            to={{
-              pathname: "/",
-            }}
+          <Button
+            variant="dark"
+            style={{ marginLeft: "30px" }}
+            onClick={this.submitForm}
           >
-            <Button
-              variant="dark"
-              style={{ marginLeft: "30px" }}
-              onClick={this.submitForm}
-            >
-              Login
-            </Button>
-          </Link>
+            Login
+          </Button>
         </Form>
       </Container>
     );
