+++ frontend/src/components/Authentication/SignupView.js
@@ -34,7 +34,7 @@ class SignupView extends Component {
       email: this.state.email,
     };
 
-    if (this.state.username.length === 0 || this.state.password.length === 0) {
+    if (this.state.username.length == 0 || this.state.password.length == 0) {
       alert("Username or password field is empty.");
       return;
     }
