+++ package.json
@@ -30,6 +30,7 @@
     "mysql": "^2.18.1",
     "nodemon": "^2.0.7",
     "react-map-gl-geocoder": "^2.1.6",
+    "react-notifications": "^1.7.2",
     "socket.io": "^4.0.0"
   }
 }
